airflow-dbt-transformation/requirements.txt
apache-airflow-providers-snowflake
astronomer-cosmos
dbt-snowflake
